+++ test/experimental/features/new_trigger_filters/filters.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2022 The Knative Authors
+Copyright 2023 The Knative Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -32,177 +32,6 @@ import (
 	"knative.dev/eventing/test/rekt/resources/broker"
 	"knative.dev/eventing/test/rekt/resources/trigger"
 )
-
-// FiltersFeatureSet creates a feature set for testing the broker implementation of the new trigger filters experimental feature
-// (aka Cloud Events Subscriptions API filters). It requires a created and ready Broker resource with brokerName.
-//
-// The feature set tests four filter dialects: exact, prefix, suffix and cesql (aka CloudEvents SQL).
-func FiltersFeatureSet(brokerName string) *feature.FeatureSet {
-	matchedEvent := FullEvent()
-	unmatchedEvent := MinEvent()
-	unmatchedEvent.SetType("org.wrong.type")
-	unmatchedEvent.SetSource("org.wrong.source")
-
-	features := make([]*feature.Feature, 0, 8)
-	tests := map[string]struct {
-		filters []eventingv1.SubscriptionsAPIFilter
-		step    feature.StepFn
-	}{
-		"Exact filter": {
-			filters: []eventingv1.SubscriptionsAPIFilter{
-				{
-					Exact: map[string]string{
-						"type":   matchedEvent.Type(),
-						"source": matchedEvent.Source(),
-					},
-				},
-			},
-		},
-		"Prefix filter": {
-			filters: []eventingv1.SubscriptionsAPIFilter{
-				{
-					Prefix: map[string]string{
-						"type":   matchedEvent.Type()[:4],
-						"source": matchedEvent.Source()[:4],
-					},
-				},
-			},
-		},
-		"Suffix filter": {
-			filters: []eventingv1.SubscriptionsAPIFilter{
-				{
-					Suffix: map[string]string{
-						"type":   matchedEvent.Type()[5:],
-						"source": matchedEvent.Source()[5:],
-					},
-				},
-			},
-		},
-		"CloudEvents SQL filter": {
-			filters: []eventingv1.SubscriptionsAPIFilter{
-				{
-					CESQL: fmt.Sprintf("type = '%s' AND source = '%s'", matchedEvent.Type(), matchedEvent.Source()),
-				},
-			},
-		},
-	}
-
-	for name, fs := range tests {
-		matchedSender := feature.MakeRandomK8sName("sender")
-		unmatchedSender := feature.MakeRandomK8sName("sender")
-		subscriber := feature.MakeRandomK8sName("subscriber")
-		triggerName := feature.MakeRandomK8sName("viaTrigger")
-
-		f := feature.NewFeatureNamed(name)
-
-		f.Setup("Install trigger subscriber", eventshub.Install(subscriber, eventshub.StartReceiver))
-
-		// Set the Trigger subscriber.
-		cfg := []manifest.CfgFn{
-			trigger.WithSubscriber(service.AsKReference(subscriber), ""),
-			trigger.WithNewFilters(fs.filters),
-		}
-
-		f.Setup("Install trigger", trigger.Install(triggerName, brokerName, cfg...))
-		f.Setup("Wait for trigger to become ready", trigger.IsReady(triggerName))
-		f.Setup("Broker is addressable", k8s.IsAddressable(broker.GVR(), brokerName))
-
-		f.Requirement("Install matched event sender", eventshub.Install(matchedSender,
-			eventshub.StartSenderToResource(broker.GVR(), brokerName),
-			eventshub.InputEvent(matchedEvent)),
-		)
-
-		f.Requirement("Install unmatched event sender", eventshub.Install(unmatchedSender,
-			eventshub.StartSenderToResource(broker.GVR(), brokerName),
-			eventshub.InputEvent(unmatchedEvent)),
-		)
-
-		f.Alpha("Triggers with new filters").
-			Must("must deliver matched events", OnStore(subscriber).MatchEvent(HasId(matchedEvent.ID())).AtLeast(1)).
-			MustNot("must not deliver unmatched events", OnStore(subscriber).MatchEvent(HasId(unmatchedEvent.ID())).Not())
-		features = append(features, f)
-	}
-
-	return &feature.FeatureSet{
-		Name:     "New trigger filters",
-		Features: features,
-	}
-}
-
-type CloudEventsContext struct {
-	eventType            string
-	eventSource          string
-	eventSubject         string
-	eventID              string
-	eventDataSchema      string
-	eventDataContentType string
-	shouldDeliver        bool
-}
-
-func AnyFilterFeature(brokerName string) *feature.Feature {
-	f := feature.NewFeature()
-
-	eventContexts := []CloudEventsContext{
-		{
-			eventType:     "exact.event.type",
-			shouldDeliver: true,
-		},
-		{
-			eventType:     "prefix.event.type",
-			shouldDeliver: true,
-		},
-		{
-			eventType:     "event.type.suffix",
-			shouldDeliver: true,
-		},
-		{
-			eventType:     "not.type.event",
-			shouldDeliver: true,
-		},
-		{
-			eventType:     "cesql.event.type",
-			shouldDeliver: true,
-		},
-		{
-			eventType:     "not.event.type",
-			shouldDeliver: false,
-		},
-	}
-
-	filters := []eventingv1.SubscriptionsAPIFilter{
-		{
-			Any: []eventingv1.SubscriptionsAPIFilter{
-				{
-					Exact: map[string]string{
-						"type": "exact.event.type",
-					},
-				},
-				{
-					Prefix: map[string]string{
-						"type": "prefix",
-					},
-				},
-				{
-					Suffix: map[string]string{
-						"type": "suffix",
-					},
-				},
-				{
-					Not: &eventingv1.SubscriptionsAPIFilter{
-						CESQL: "type LIKE %event.type%",
-					},
-				},
-				{
-					CESQL: "type = 'cesql.event.type'",
-				},
-			},
-		},
-	}
-
-	f = newEventFilterFeature(eventContexts, filters, f, brokerName)
-
-	return f
-}
 
 func newEventFilterFeature(eventContexts []CloudEventsContext, filters []eventingv1.SubscriptionsAPIFilter, f *feature.Feature, brokerName string) *feature.Feature {
 	subscriberName := feature.MakeRandomK8sName("subscriber")
