+++ test/experimental/new_trigger_filters_test.go
@@ -49,6 +49,22 @@ func TestMTChannelBrokerNewTriggerFilters(t *testing.T) {
 	env.TestSet(ctx, t, newfilters.FiltersFeatureSet(brokerName))
 }
 
+func TestMTChannelBrokerAnyTriggerFilters(t *testing.T) {
+	t.Parallel()
+
+	ctx, env := global.Environment(
+		knative.WithKnativeNamespace(system.Namespace()),
+		knative.WithLoggingConfig,
+		knative.WithTracingConfig,
+		k8s.WithEventListener,
+		environment.Managed(t),
+	)
+	brokerName := "default"
+
+	env.Prerequisite(ctx, t, InstallMTBroker(brokerName))
+	env.Test(ctx, t, newfilters.AnyFilterFeature(brokerName))
+}
+
 func InstallMTBroker(name string) *feature.Feature {
 	f := feature.NewFeatureNamed("Multi-tenant channel-based broker")
 	f.Setup(fmt.Sprintf("Install broker %q", name), broker.Install(name, broker.WithEnvConfig()...))
